Refactor sidebar component in dashboard page
