+++ src/app/main/dashboard2/page.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/app-sidebar"
+import { AppSidebar } from "@/components/AppSidebar"
 import Charts from "@/components/charts-01"
 import { DashboardPreview } from "@/components/dashboard-preview"
 import { OrganizationFormComponent } from "@/components/organization-form"
@@ -13,7 +13,7 @@ export default async function Page() {
   return (
 
     <TooltipProvider>
-          <SidebarTrigger />
+      <SidebarTrigger />
       <main className="flex flex-1 flex-col p-2 transition-all duration-300 ease-in-out">
         <DashboardPreview />
       </main>
