+++ src/app/main/organization/page.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/app-sidebar"
+import { AppSidebar } from "@/components/AppSidebar"
 import { OrganizationFormComponent } from "@/components/organization-form"
 import {
   SidebarLayout,
@@ -8,16 +8,16 @@ import {
 export default function OrganizationPage() {
   return (
 
-      <main className="flex-1 overflow-y-auto">
-              <SidebarTrigger />
+    <main className="flex-1 overflow-y-auto">
+      <SidebarTrigger />
 
-        <div className="container mx-auto py-8">
-          <h1 className="text-2xl font-bold mb-4 responsive-bg smobile:responsive-bg sm:responsive-bg md:responsive-bg lg:responsive-bg xl:responsive-bg">
-            Organization List
-          </h1>
-          <OrganizationFormComponent />
-        </div>
-      </main>
+      <div className="container mx-auto py-8">
+        <h1 className="text-2xl font-bold mb-4 responsive-bg smobile:responsive-bg sm:responsive-bg md:responsive-bg lg:responsive-bg xl:responsive-bg">
+          Organization List
+        </h1>
+        <OrganizationFormComponent />
+      </div>
+    </main>
 
   )
 }
