+++ src/components/AppSidebar.tsx
@@ -249,7 +249,7 @@ const data = {
 
 export function AppSidebar() {
   return (
-    <Sidebar className="mt-[50px] flex flex-col h-[calc(100vh-50px)]">
+    <Sidebar className="mt-[50px] flex flex-col h-[calc(100vh-50px)] bg-inherit">
       <SidebarHeader>
         <TeamSwitcher teams={data.teams} />
       </SidebarHeader>
