+++ src/components/TopMenuBar.tsx
@@ -1,3 +1,4 @@
+"use client"
 import type { Metadata } from "next";
 
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
@@ -5,6 +6,7 @@ import { Button } from "@/components/ui/button";
 import { Link, ChevronDown } from "lucide-react";
 import React from "react";
 import { ModeToggle } from "./ModeToggle";
+import { useTheme } from "next-themes";
 
 export const metadata: Metadata = {
   title: "Create Next App",
@@ -12,18 +14,20 @@ export const metadata: Metadata = {
 };
 
 export default function TopMenuBar() {
+  const { theme } = useTheme()
+
   return (
-    <header className="fixed top-0 left-0 right-0 z-50 bg-black border-b">
+    <header className="fixed top-0 left-0 right-0 z-50 border-b">
       <div className="w-full flex  justify-between h-12 px-4">
         <div className="flex items-center">
           <span className="flex items-center mr-4">
             <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="mr-2"><path d="M12 14a2 2 0 1 0 0-4 2 2 0 0 0 0 4z" /><path d="M20 14a2 2 0 1 0 0-4 2 2 0 0 0 0 4z" /><path d="M4 14a2 2 0 1 0 0-4 2 2 0 0 0 0 4z" /><path d="M12 8v-2" /><path d="M12 18v-2" /><path d="M18.7 8.7l-1.4-1.4" /><path d="M6.7 6.7l-1.4-1.4" /><path d="M18.7 15.3l-1.4 1.4" /><path d="M6.7 17.3l-1.4 1.4" /></svg>
-            <span className="font-bold text-xl tracking-tight font-logo text-white">Demo App</span>
+            <span className="font-bold text-xl tracking-tight font-logo">Demo App</span>
           </span>
           <nav className="hidden md:flex items-center space-x-1">
             <DropdownMenu>
               <DropdownMenuTrigger asChild>
-                <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Products <ChevronDown className="ml-1 h-3 w-3" /></Button>
+                <Button variant="ghost" size="default" className=" hover:bg-gray-800">Products <ChevronDown className="ml-1 h-3 w-3" /></Button>
               </DropdownMenuTrigger>
               <DropdownMenuContent>
                 <DropdownMenuItem>Product 1</DropdownMenuItem>
@@ -32,17 +36,17 @@ export default function TopMenuBar() {
             </DropdownMenu>
             <DropdownMenu>
               <DropdownMenuTrigger asChild>
-                <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Solutions <ChevronDown className="ml-1 h-3 w-3" /></Button>
+                <Button variant="ghost" size="default" className="hover:bg-gray-800">Solutions <ChevronDown className="ml-1 h-3 w-3" /></Button>
               </DropdownMenuTrigger>
               <DropdownMenuContent>
                 <DropdownMenuItem>Solution 1</DropdownMenuItem>
                 <DropdownMenuItem>Solution 2</DropdownMenuItem>
               </DropdownMenuContent>
             </DropdownMenu>
-            <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Resources</Button>
-            <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Enterprise</Button>
-            <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Docs</Button>
-            <Button variant="ghost" size="sm" className="text-white hover:bg-gray-800">Pricing</Button>
+            <Button variant="ghost" size="default" className="hover:bg-gray-800">Resources</Button>
+            <Button variant="ghost" size="default" className="hover:bg-gray-800">Enterprise</Button>
+            <Button variant="ghost" size="default" className="hover:bg-gray-800">Docs</Button>
+            <Button variant="ghost" size="default" className="hover:bg-gray-800">Pricing</Button>
           </nav>
         </div>
         <div className="flex items-center space-x-2">
